debug client

+++ source/server/include/protocol.hpp
@@ -21,7 +21,7 @@ namespace server {
 	inline opcode operator|(opcode lh, opcode rh){ return static_cast<opcode>((uint8_t)lh | rh); }
 	inline opcode operator&(opcode lh, opcode rh){ return static_cast<opcode>((uint8_t)lh & rh); }
 	inline opcode operator^(opcode lh, opcode rh){ return static_cast<opcode>((uint8_t)lh ^ rh); }
-	inline opcode operator~(opcode lh){ return static_cast<opcode>(~lh); }
+	inline opcode operator~(opcode lh){ return static_cast<opcode>(~(uint8_t)lh); }
 }
